SSCS-3581 Submit answers to a question
When you enter an answer to a question you can either save it as draft
or submit it. If you submit you answer initially an changes you made to
your answer will be saved as draft and you will be presented with a new
screen asking you to confirm you want to submit you answer. If you
select ok the question will be submitted and you can no longer change.
The user is then taken back to the task list page.