SSCS-4349 Allow JS evidence upload errors
With JS enabled if the first file you uploaded was too big or not of
an acceptable type the page would reload but the evidence upload
checkbox was not ticked and the error message was hidden. Make
evidence upload section visable if you have uploaded files or have
an errors.