SSCS-4412 Remember updates to answer when uploading evidence
When you upload or delete evidence with JS enabled any updates to your
answer were being lost. Save the updates as draft every time we delete
or add evidence. This meant we had to change the way delete works.
Before we had a separate form for each delete but changes this to be a
single form and encode the evidence id into the delete button name.