Remove git dependency
We are relying on a bug fix to Pa11y that has not been merged into their
project yet. This is hosted on a fork of Pa11y. Previously the
dependency in package.json was of the form git+http://github.com...
This did not work when we tried to build our app on the docker image
provided to CNP as it did not have git installed. We were doing an
apt-get update && apt-get install git every time. Changed the dependency
to use the tarball from github created by a tag which does not need git.