SSCS-4424 Changed links to secondary buttons
User research has shown that the links to save you answer on the answer
page or go back to question on the submit confirmation page are not
being seen by users. Changed them to secondary buttons to make them
more visible. Also removed some logs in login as I think they will be
writing user details to the logs.