SSCS-4093 Handle cases where we cannot load an appeal
When you login to cor there are a number of reasons we cannot load your
appeal.

- User creates an account, logs in, but there is no case in CCD. A
subcase of this is user has been participating in COH, changes their
subscription email address and then logs in using their old email
address.
- User creates an account, logs in, there is a case in CCD but it's not
marked for COR - active appeal but not marked for COR
- User logs in and there are multiple cases associated with the email
address - email for support
- User creates an account, logs in, there is a case, it is marked for
COH but no questions have have asked yet

Added code to whilst logging in to handle the first 3 cases and to
display the task list with no questions for the last.