Allow us to set a case id when we login
If you have an email address with multiple cases you can normally you
cannot login as we cannot decide which case to show. Allow you to edit
url you are taken to once you cannot login to add a query parameter
caseId that will take you to that case if it is associated with the
current user.

e.g. http://{domain}/sign-in?code={code}&caseId={case_id}