SSCS-5230 Show final decision
When the final decision page was originally written we expected the COH
state to reflect the final decision. This has not been done so we have
added has_final_decision to the api call on the backend when loading a
hearing. This looks in the the appeals history to see if a final
decision has been issued. Use this to decide if we show the final
decision page.