SSCS-5230 View final decision when there is no Preliminary View
In JUI the panel can issue a final decision without issuing a
preliminary view. COR could not handle this as it tries to parse
it as a Preliminary view and some of the data is not there. Update
routing to use the has_final_decision attribute of the hearing to
display the final decision screen.