SSCS-3973 Decision title missing
When we wrote this there was an idea that the appellant would see an
accepted or rejected decision. But if the panel rejects the appeal they
will relist it rather than show the appeal rejected here. Changed the
title of the decision page to be static as it no longer needs to
select between accepted and rejected and the states are not set anyway.