SSCS-5808 Pass tya number of state to idam
The links sent to appellants will contain a tya number. When the user
clicks the link we will initially have the tya number in the url but
then we redirect them to Idam. Idam has a mechanism to preserve values
where you add a state parameter to the request sent to idam. Then on the
callback after the user has logged in we get given the state parameter
back. The user flow is as follows,

- User clicks link contains tya number
- redirect to idam add tya number as state parameter
- user registers on idam
- user is redirected back to service but is not logged in. Get state
parameter and redirect back to idam for login including state.
- user logins into idam
- user is redirected back to service and is logged in with state
parameter.

Currently just save the state parameter in the session but will need to
use it to associate the case with a sidam account.