SSCS-5808 Assign case to user
When a user logs in they either
- have no cases assigned to them so we need to assign the case.
- have a case assigned to them so we can show the case.
- have multiple cases assigned to them so we need to select the case to
show them.

Added logic to the login controller to load which cases the citizen has
assigned to them and if need be assign the case to them. If a tya number
has been provided then the backend service will filter out cases that
match the tya number.