SSCS-6134 Load MYA even if an SC number has been set
Currently MYA fails with an error if the case does not have an SC
number. This is an edge case as the appellant should not have got a link
to MYA until their case has an SC number but sometimes we will send one
out. The only issue was we spell out the sc number for screen readers
so made that function handle undefined and display 'to be confirmed' on
the page.