SSCS-6282 Add feature flag to turn on bulk scan coversheet
Some pages talk about posting in evidence. At the moment these should
all ask you to send it in and write the sc number on it. In future we
will send them to bulk scanning and give the user the option to download
a cover sheet. Added a feature flag postBulkScan which is off by default
to enable the coversheet page.