SSCS-6334 Allow non COR users to use MYA
To upload evidence or submit a statement you needed to have an
online hearing. This was becase the backend api used the online
hearing id. This has now been changed so it can use the online
hearing id or the CCD case id. Made the changes in the frontend
to always use the CCD case id.