SSCS-6662 Display subscription email and sms details (#404)
* SSCS-6662 Display subscription email and sms details

We were showing a single email and sms address on the your-details page.
This is not really true as you could have a number of subscriptions. As
MYA does not know what type of user has logged in (appellant, appointee,
representative) we show all the subscriptions a case has on the your
details page.